requirements.txt
Django==1.11.7
django-webpack-loader==0.5.0
djangorestframework==3.7.1
model-mommy==1.4.0